pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zerozen"
version = "2025.8.10"
description = "hyper-personal, always-on, open-source AI companion."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.7.1",
    "rich>=14.0.0",
    "typer>=0.16.0",
    "python-dotenv>=1.0.0",
    "openai>=1.97.0",
    "openai-agents>=0.2.2",
    "google-api-python-client>=2.178.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "backoff>=2.2.1",
    "google-api-core>=2.25.1",
    "google-auth>=2.40.3",
    "googleapis-common-protos>=1.70.0",
    "sqlmodel>=0.0.22",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "chromadb>=0.4.22",
]

[project.scripts]
zen="zerozen.cli:app"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "coverage>=7.10.2",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.12.4",
]
